Initial commit
We expliciti the contract we want to implement through the project.
This means that we want people to use the Fraction module with functions
like:

- Fraction.print/1
- Fraction.add/2
- Fraction.proper?/1

And so on. The internal details should not be visible to the outside
world. People don't need to know how we do the calculations under the hood.